Remove searchParams to opt out of dynamic render
According to the Next.js documentation, the
searchParams Pages propr opts a route into
dynamic rendering. What I want to achieve is
render the top-level page statically and fetch
the initial data on the server, then pass that
data to the client. Thereafter, the client should
not ask for the top-level Page component again,
which would thereby do the data fetching again.

See: https://nextjs.org/docs/app/building-your-application/rendering/server-components#dynamic-functions

To achieve it, we do what we were already doing:
rely on the useSearchParams hook to get the query
parameters we need.