Seed PokeAPI data into MongoDB database
The PokeAPI rules say that requested data must be fetched
locally. To help abide by this rule, the seed script fetches
the data that we expect to use in the App. Later, we will make
the MongoDB database the source of truth for the pokemon data
used by the application.