Fetch all pokesearch from Mongo on initial load
The roto Page component in the /pokesearch route
now queries the MongoDB database for all Pokemon
instead of making a request to PokeAPI. Moreover,
the requested was augmented to not only include
the name of each Pokemon, but also everything else
necessary to make a card. As a result, the client
no longer needs to make requests when paginating,
since it has all the data to display the cards.